requirements.txt
fastapi==0.78.0
keras==2.10.0
keras_bert==0.88.0
MatchZoo==2.2.0
numpy==1.20.0
tensorflow==2.10.0
pydantic==1.9.0
python-dotenv==0.20.0
requests==2.26.0
uvicorn==0.17.6
